refactor(frontend): move complex hook logics to files
