+++ frontend/src/app/App/index.js
@@ -1,16 +1,13 @@
-import React, {
-  createContext,
-  useEffect,
-  useState,
-  useCallback,
-  useMemo,
-} from 'react'
+import React, { createContext, useState } from 'react'
 import styled from 'styled-components/macro'
 
 import Header from 'app//Header'
 import RecipeList from 'app/RecipeList'
 import SavedRecipeList from 'app/SavedRecipeList'
 
+import useFilters from 'utils/useFilters'
+import useRecipes from 'utils/useRecipes'
+import useSavedRecipes from 'utils/useSavedRecipes'
 import * as api from 'utils/api'
 
 const Layout = styled.div`
@@ -27,177 +24,51 @@ const Layout = styled.div`
 export const AppContext = createContext()
 
 const App = () => {
-  const [user, setUser] = useState()
-  useEffect(() => {
-    const fetchData = async () => {
-      const user = await api.getUser()
-      setUser(user)
-    }
-
-    fetchData()
-  }, [])
-
-  const [filters, setFilter] = useState({
-    query: '',
-    liked: 0,
-  })
-  const addFilter = (label, value) => {
-    setFilter((filters) => ({
-      ...filters,
-      [label]: value,
-    }))
-    setOffset(0)
-  }
-
-  const [canLoadMore, setCanLoadMore] = useState(true)
-  const checkIfCanLoadMore = (items) => {
-    setCanLoadMore(items.length === api.OFFSET)
-  }
-
-  const [recipes, setRecipes] = useState([])
   const [offset, setOffset] = useState(0)
-  const [hasLoadedRecipes, setHasLoadedRecipes] = useState(false)
-  const [isLoadingRecipes, setIsLoadingRecipes] = useState(false)
   const [sortBy, setSortBy] = useState(api.SORT_BY.CREATED_DESC)
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoadingRecipes(true)
-      const newRecipes = await api.getRecipes({
-        keywords: filters.query,
-        sortBy,
-        liked: filters.liked,
-      })
-      setRecipes(newRecipes)
-      setIsLoadingRecipes(false)
-      setHasLoadedRecipes(true)
 
-      setOffset(0)
-      checkIfCanLoadMore(newRecipes)
-    }
+  const { filters, addFilter } = useFilters(
+    {
+      query: '',
+      liked: 0,
+    },
+    { setOffset },
+  )
 
-    const timeout = setTimeout(fetchData, 300)
-    return () => clearTimeout(timeout)
-  }, [filters.liked, filters.query, sortBy])
+  const {
+    recipes,
+    setRecipes,
+    hasLoadedRecipes,
+    isLoadingRecipes,
+    canLoadMoreRecipes,
+    loadMoreRecipes,
+    toggleLikeRecipe,
+    excludeRecipe,
+  } = useRecipes([], {
+    filters,
+    sortBy,
+    offset,
+    setOffset,
+  })
 
-  const loadMore = async () => {
-    const newOffset = offset + api.OFFSET
-    setIsLoadingRecipes(true)
-    const newRecipes = await api.getRecipes({
-      keywords: filters.query,
-      offset: newOffset,
-      sortBy,
-      liked: filters.liked,
-    })
-    setRecipes(recipes.concat(newRecipes))
-    setIsLoadingRecipes(false)
+  const {
+    savedRecipes,
+    hasLoadedSavecRecipes,
+    isLoadingSavecRecipes,
+    toggleSaveRecipe,
+    generateSavedRecipes,
+  } = useSavedRecipes([], { setRecipes })
 
-    setOffset(newOffset)
-    checkIfCanLoadMore(newRecipes)
+  const contextValue = {
+    toggleSaveRecipe,
+    toggleLikeRecipe,
+    excludeRecipe,
   }
-
-  const [savedRecipes, setSavedRecipes] = useState([])
-  const [hasLoadedSavecRecipes, setHasLoadedSavecRecipes] = useState(false)
-  const [isLoadingSavecRecipes, setIsLoadingSavecRecipes] = useState(false)
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoadingSavecRecipes(true)
-      setSavedRecipes(await api.getSavedRecipes())
-      setIsLoadingSavecRecipes(false)
-      setHasLoadedSavecRecipes(true)
-    }
-
-    fetchData()
-  }, [])
-
-  const generateRecipes = useCallback(async () => {
-    const savedRecipes = await api.generateRecipes()
-    setSavedRecipes(savedRecipes)
-    setRecipes((recipes) =>
-      recipes.map((recipe) => ({
-        ...recipe,
-        saved: savedRecipes.some((savedRecipe) => savedRecipe.id === recipe.id),
-      })),
-    )
-  }, [])
-
-  const toggleSaveRecipe = useCallback(
-    (recipe) => {
-      const isRecipeSaved = savedRecipes.some((r) => r.id === recipe.id)
-      if (isRecipeSaved) {
-        api.unsaveRecipe(recipe.id)
-        setSavedRecipes(savedRecipes.filter(({ id }) => id !== recipe.id))
-      } else {
-        api.saveRecipe(recipe.id)
-        setSavedRecipes([
-          ...savedRecipes,
-          {
-            ...recipe,
-            saved: true,
-          },
-        ])
-      }
-
-      setRecipes((recipes) =>
-        recipes.map((r) => {
-          if (r.id !== recipe.id) return r
-          return {
-            ...recipe,
-            saved: !isRecipeSaved,
-          }
-        }),
-      )
-    },
-    [savedRecipes],
-  )
-
-  const toggleLikeRecipe = useCallback(
-    (recipe) => {
-      if (recipe.liked) {
-        api.unlikeRecipe(recipe.id)
-      } else {
-        api.likeRecipe(recipe.id)
-      }
-
-      setRecipes(
-        recipes.map((r) => {
-          if (r.id !== recipe.id) return r
-          return {
-            ...recipe,
-            liked: !recipe.liked,
-          }
-        }),
-      )
-    },
-    [recipes],
-  )
-
-  const excludeRecipe = useCallback(
-    (recipe) => {
-      if (
-        window.confirm(
-          'Do you really want to exclude this recipe? It will be hidden for ever ever',
-        )
-      ) {
-        api.excludeRecipe(recipe.id)
-        setRecipes(recipes.filter((r) => r.id !== recipe.id))
-      }
-    },
-    [recipes],
-  )
-
-  const contextValue = useMemo(
-    () => ({
-      toggleSaveRecipe,
-      toggleLikeRecipe,
-      excludeRecipe,
-    }),
-    [toggleSaveRecipe, toggleLikeRecipe, excludeRecipe],
-  )
 
   return (
     <AppContext.Provider value={contextValue}>
       <Layout>
-        <Header filters={filters} addFilter={addFilter} user={user} />
+        <Header filters={filters} addFilter={addFilter} />
         <RecipeList
           hasLoaded={hasLoadedRecipes}
           isLoading={isLoadingRecipes}
@@ -205,8 +76,8 @@ const App = () => {
           toggleSaveRecipe={toggleSaveRecipe}
           toggleLikeRecipe={toggleLikeRecipe}
           excludeRecipe={excludeRecipe}
-          canLoadMore={canLoadMore}
-          loadMore={loadMore}
+          canLoadMoreRecipes={canLoadMoreRecipes}
+          loadMore={loadMoreRecipes}
           sortBy={sortBy}
           onChangeSortBy={setSortBy}
         />
@@ -215,7 +86,7 @@ const App = () => {
           isLoading={isLoadingSavecRecipes}
           savedRecipes={savedRecipes}
           removeSavedRecipe={toggleSaveRecipe}
-          generateRecipes={generateRecipes}
+          generateSavedRecipes={generateSavedRecipes}
         />
       </Layout>
     </AppContext.Provider>
